test: add comprehensive tests and improve robustness for relay_setup.py
- Refactored tests/test_relay_setup.py to follow "mirror source modules" rule.
- Moved Settings.from_relay_config tests to tests/test_config.py.
- Moved lifespan integration tests to tests/test_server.py.
- Added OSError handling to check_saved_sessions in relay_setup.py.
- Added edge case tests for check_saved_sessions and _sanitize_error.
- Achieved 100% test coverage for relay_setup.py.
- Updated tests/test_server.py to be resilient to optional tools.
